backend/requirements.txt
# Core scientific packages
numpy>=1.20.0
pandas>=1.3.0

# Chemistry tools
rdkit>=2022.03.1

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Visualization and structure handling
biopython==1.81
meeko>=0.4.0
vina>=1.2.0
py3Dmol>=1.8.0

loguru==0.7.2
requests==2.31.0

# OpenBabel - We use the wheel package only and remove direct openbabel
openbabel-wheel==3.1.1.21  # Pre-built wheel for OpenBabel

# Web framework (FastAPI instead of Flask)
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5  # For handling file uploads

# Other utilities
tqdm>=4.62.0
scikit-learn>=0.24.0
pydantic>=1.8.0

reportlab>=4.4.0